requirements.txt
torch
torchvision
yaml
pillow
